moved check mode ckeck to format_file_source
